% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.caret.R
\docType{data}
\name{catboost.caret}
\alias{catboost.caret}
\title{Support caret interface}
\format{
An object of class \code{list} of length 10.
}
\usage{
catboost.caret
}
\description{
Support caret interface
}
\keyword{datasets}
