% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{catboost.cv}
\alias{catboost.cv}
\title{Cross-validate model.}
\usage{
catboost.cv(pool, params = list(), fold_count = 3, inverted = FALSE,
  partition_random_seed = 0, shuffle = TRUE, stratified = FALSE,
  early_stopping_rounds = NULL)
}
\arguments{
\item{pool}{Data to cross-validatte}

\item{params}{Parameters for catboost.train}

\item{fold_count}{Folds count.}

\item{inverted}{Train on the test fold and evaluate the model on the training folds.}

\item{partition_random_seed}{The random seed used for splittng pool into folds.}

\item{shuffle}{Shuffle the dataset objects before splitting into folds.}

\item{stratified}{Perform stratified sampling.}

\item{early_stopping_rounds}{Activates Iter overfitting detector with od_wait set to early_stopping_rounds.}

\item{pool}{The dataset used for cross-validation.}
}
\description{
Cross-validate model.
}
