% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{catboost.drop_unused_features}
\alias{catboost.drop_unused_features}
\title{Drop unused features information from model}
\usage{
catboost.drop_unused_features(model, ntree_end, ntree_start = 0)
}
\arguments{
\item{model}{The model obtained as the result of training.}

\item{ntree_end}{Leave the trees with indices from the interval [ntree_start, ntree_end) (zero-based indexing).}

\item{ntree_start}{Leave the trees with indices from the interval [ntree_start, ntree_end) (zero-based indexing).

Default value: 0}
}
\description{
Drop unused features information from model
}
