% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{catboost.eval_metrics}
\alias{catboost.eval_metrics}
\title{Calculate metrics.}
\usage{
catboost.eval_metrics(
  model,
  pool,
  metrics,
  ntree_start = 0L,
  ntree_end = 0L,
  eval_period = 1,
  thread_count = -1,
  tmp_dir = NULL
)
}
\arguments{
\item{model}{The model obtained as the result of training.

Default value: Required argument}

\item{pool}{The pool for which you want to evaluate the metrics.

Default value: Required argument}

\item{metrics}{The list of metrics to be calculated.
(Supported metrics https://catboost.ai/docs/references/custom-metric__supported-metrics.html)

Default value: Required argument}

\item{ntree_start}{Model is applied on the interval [ntree_start, ntree_end) with the step eval_period (zero-based indexing).

Default value: 0}

\item{ntree_end}{Model is applied on the interval [ntree_start, ntree_end) with the step eval_period (zero-based indexing).

Default value: 0 (if value equals to 0 this parameter is ignored and ntree_end equal to tree_count)}

\item{eval_period}{Model is applied on the interval [ntree_start, ntree_end) with the step eval_period (zero-based indexing).

Default value: 1}

\item{thread_count}{The number of threads to use when applying the model.
If -1, then the number of threads is set to the number of CPU cores.

Allows you to optimize the speed of execution. This parameter doesn't affect results.

Default value: -1}

\item{tmp_dir}{The name of the temporary directory for intermediate results.
If NULL, then the name will be generated.

Default value: NULL}
}
\value{
dict: metric -> array of shape [(ntree_end - ntree_start) / eval_period].
}
\description{
Calculate the specified metrics for the specified dataset.
}
\seealso{
\url{https://catboost.ai/docs/concepts/python-reference_catboost_eval-metrics.html}
}
