% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{catboost.from_data_frame}
\alias{catboost.from_data_frame}
\title{Create pool from data frame}
\usage{
catboost.from_data_frame(data, target = NULL, weight = NULL,
  baseline = NULL)
}
\arguments{
\item{data}{data.frame object}

\item{target}{Target vector}
}
\description{
Create pool from data frame
}
