% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{catboost.from_file}
\alias{catboost.from_file}
\title{catboost.from_file}
\usage{
catboost.from_file(pool_path, cd_path = "", pairs_path = "",
  delimiter = "\\t", has_header = FALSE, thread_count = 1,
  verbose = FALSE)
}
\arguments{
\item{pool_path}{The path to the input file that contains the dataset description.

Default value: Required argument}

\item{cd_path}{The path to the input file that contains the column descriptions.}

\item{pairs_path}{The path to the input file that contains the pairs descriptions.}

\item{delimiter}{Delimiter character to use to separate features in a file.}

\item{has_header}{Read column names from first line, if this parameter is set to True.}

\item{thread_count}{The number of threads to use while reading the data. Optimizes reading time. This parameter doesn't affect results.}

\item{verbose}{Verbose output to stdout.}
}
\value{
catboost.Pool
}
\description{
Create a dataset from the given file.
}
\examples{
pool_path <- system.file("extdata", "adult_train.1000", package = "catboost")
test_pool_path <- system.file("extdata", "adult_test.1000", package = "catboost")
cd_path <- system.file("extdata", "adult.cd", package = "catboost")
pool <- catboost.from_file(pool_path, cd_path)
test_pool <- catboost.from_file(test_pool_path, cd_path)

}
