% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{catboost.get_feature_importance}
\alias{catboost.get_feature_importance}
\title{Calculate the feature importances}
\usage{
catboost.get_feature_importance(model, pool = NULL,
  type = "PredictionValuesChange", thread_count = -1)
}
\arguments{
\item{model}{The model obtained as the result of training.

Default value: Required argument}

\item{pool}{The input dataset.

The feature importance for the training dataset is calculated if this argument is not specified.

Default value: NULL}

\item{fstr_type}{The fstr type.

Possible values:
\itemize{
  \item 'PredictionValuesChange'

    Calculate score for every feature by values change.

  \item 'LossFunctionChange'

    Calculate score for every feature bu loss change. Use it for ranking methods.

  \item 'Interaction'

    Calculate pairwise score between every feature.

  \item 'ShapValues'

    Calculate SHAP Values for every object.

}

Default value: 'PredictionValuesChange'}

\item{thread_count}{The number of threads to use when applying the model. If -1, then the number of threads is set to the number of cores.

Allows you to optimize the speed of execution. This parameter doesn't affect results.

Default value: -1}
}
\description{
Calculate the feature importances (see \url{https://tech.yandex.com/catboost/doc/dg/concepts/fstr-docpage/#fstr})
(Regular feature importance, ShapValues, and Feature interaction strength).
}
\seealso{
\url{https://tech.yandex.com/catboost/doc/dg/features/feature-importances-calculation-docpage}
}
