% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{catboost.get_model_params}
\alias{catboost.get_model_params}
\title{Model parameters}
\usage{
catboost.get_model_params(model)
}
\arguments{
\item{model}{The model obtained as the result of training.

Default value: Required argument}
}
\description{
Return the model parameters.
}
\seealso{
\url{https://catboost.ai/docs/concepts/r-reference_catboost-get_model_params.html}
}
