% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{catboost.get_plain_params}
\alias{catboost.get_plain_params}
\title{Plain Model parameters}
\usage{
catboost.get_plain_params(model)
}
\arguments{
\item{model}{he model obtained as the result of training.}
}
\value{
A list object with model parameters.
}
\description{
Return the plain model parameters.
}
