% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{catboost.importance}
\alias{catboost.importance}
\title{Estimate feature importance}
\usage{
catboost.importance(model, pool = NULL, thread_count = 1)
}
\arguments{
\item{model}{Trained model}

\item{pool}{Pool to estimate importance on}

\item{thread_count}{Number of parallel threads}
}
\description{
Estimate feature importance
}
