% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{catboost.load_pool}
\alias{catboost.load_pool}
\title{Load the catboost pool}
\usage{
catboost.load_pool(pool_path, column_description_path = "",
  thread_count = 1, verbose = FALSE)
}
\arguments{
\item{pool_path}{A path to the catboost pool file (First 4 columns has special meaning. Colums after 5-th contain features).}

\item{column_description_path}{Feature descriptor file path.}

\item{thread_count}{Number of threads used while reading data.}

\item{verbose}{Log output.}
}
\description{
Load the catboost pool
}
