% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{catboost.ncol}
\alias{catboost.ncol}
\title{catboost.ncol}
\usage{
catboost.ncol(pool)
}
\arguments{
\item{pool}{The input dataset.}
}
\value{
The number of columns.
}
\description{
Calculate the number of columns in the dataset.
}
\seealso{
\url{https://tech.yandex.com/catboost/doc/dg/concepts/r-reference_catboost-ncol-docpage/}
}
