% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{catboost.nrow}
\alias{catboost.nrow}
\title{catboost.nrow}
\usage{
catboost.nrow(pool)
}
\arguments{
\item{pool}{The input dataset.}
}
\value{
The number of objects.
}
\description{
Calculate the number of objects in the dataset.
}
\seealso{
\url{https://tech.yandex.com/catboost/doc/dg/concepts/r-reference_catboost-nrow-docpage/}
}
