% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{catboost.ntrees}
\alias{catboost.ntrees}
\title{Number of trees in the model}
\usage{
catboost.ntrees(model)
}
\arguments{
\item{model}{The model obtained as the result of training.

Default value: Required argument}
}
\description{
Return the number of trees in the model.
}
