% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{catboost.save_pool}
\alias{catboost.save_pool}
\title{Writes data.frame in pool format}
\usage{
catboost.save_pool(data, target, weight, baseline, pool_path,
  column_description_path)
}
\arguments{
\item{data}{Features data.frame}

\item{target}{Target vector}

\item{baseline}{Baseline vector}

\item{pool_path}{Pool file path}

\item{column_description_path}{Column description file path}

\item{weights}{Weights vector}
}
\description{
Writes data.frame in pool format
}
