% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{catboost.sum_models}
\alias{catboost.sum_models}
\title{Sum models.}
\usage{
catboost.sum_models(
  models,
  weights = NULL,
  ctr_merge_policy = "IntersectingCountersAverage"
)
}
\arguments{
\item{models}{Models for the summation.

Default value: Required argument}

\item{weights}{The weights of the models.

Default value: NULL (use weight 1 for every model)}

\item{ctr_merge_policy}{The counters merging policy.
Possible values:
\itemize{
  \item 'FailIfCtrIntersects'
    Ensure that the models have zero intersecting counters
  \item 'LeaveMostDiversifiedTable'
    Use the most diversified counters by the count of unique hash values
  \item 'IntersectingCountersAverage'
    Use the average ctr counter values in the intersecting bins
}

Default value: 'IntersectingCountersAverage'}
}
\value{
Model object.
}
\description{
Blend trees and counters of two or more trained CatBoost models into a new model.
             Leaf values can be individually weighted for each input model. For example, it may
             be useful to blend models trained on different validation datasets.
}
