% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{catboost.train}
\alias{catboost.train}
\title{Train the catboost model on catboost pool}
\usage{
catboost.train(learn_pool, test_pool = NULL, params = list(),
  calc_importance = FALSE)
}
\description{
Train the catboost model on catboost pool
}
