% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{catboost.virtual_ensembles_predict}
\alias{catboost.virtual_ensembles_predict}
\title{Apply the model with several virtual ensembles}
\usage{
catboost.virtual_ensembles_predict(
  model,
  pool,
  verbose = FALSE,
  prediction_type = "VirtEnsembles",
  ntree_end = 0L,
  virtual_ensembles_count = 10,
  thread_count = -1
)
}
\arguments{
\item{model}{The model obtained as the result of training.

Default value: Required argument}

\item{pool}{The input dataset.

Default value: Required argument}

\item{verbose}{Verbose output to stdout.

Default value: FALSE (not used)}

\item{prediction_type}{The format for displaying approximated values in output data
(see \url{https://catboost.ai/docs/concepts/python-reference_virtual_ensembles_predict.html#python-reference_catboostclassifier_predict__output-format}).

Possible values:
\itemize{
  \item 'VirtEnsembles'
  \item 'TotalUncertainty'
}

Default value: 'VirtEnsembles'}

\item{ntree_end}{Index of the first tree not to be used when applying the model or calculating the metrics (zero-based indexing).

Default value: 0 (the index of the last tree to use equals to the number of trees in the model minus one)}

\item{virtual_ensembles_count}{Number of tree ensembles to use. Each virtual ensemble can be considered as truncated model.

Default value: 10}

\item{thread_count}{The number of threads to use when applying the model. If -1, then the number of threads is set to the number of CPU cores.

Allows you to optimize the speed of execution. This parameter doesn't affect results.

Default value: -1}
}
\value{
Matrix or Array of predictions (for 'TotalUncertainty' and 'VirtEnsembles' prediction_type correspondingly)
}
\description{
Apply the model to the given dataset using several independent truncated models - virtual ensembles. Each tree in
             ensemble predicts its own value for each document from pool.

             Peculiarities: Return value varies on prediction_type: array for 'VirtEnsembles' and matrix for 'TotalUncertainty'
}
\seealso{
\url{https://catboost.ai/docs/concepts/python-reference_virtual_ensembles_predict.html?lang=en}
}
