% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{dim.catboost.Pool}
\alias{dim.catboost.Pool}
\title{Dimensions of catboost.Pool}
\usage{
\method{dim}{catboost.Pool}(x)
}
\arguments{
\item{x}{The input dataset.

Default value: Required argument}
}
\description{
Returns a vector of row numbers and column numbers in an catboost.Pool.
}
