% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{dimnames.catboost.Pool}
\alias{dimnames.catboost.Pool}
\title{Dimension names of catboost.Pool}
\usage{
\method{dimnames}{catboost.Pool}(x)
}
\arguments{
\item{x}{The input dataset.

Default value: Required argument}
}
\value{
A list with the two elements. The second element contains the column names.
}
\description{
Get dimension names of a Pool.
}
