% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{head.catboost.Pool}
\alias{head.catboost.Pool}
\title{Head of catboost.Pool}
\usage{
\method{head}{catboost.Pool}(x, n = 10, ...)
}
\arguments{
\item{x}{The input dataset.

Default value: Required argument}

\item{n}{The quantity of the first objects in the dataset to be returned.

Default value: 10}

\item{...}{not currently used}
}
\value{
A matrix containing the first \code{n} objects of the dataset.
}
\description{
Return a list with the first n objects of the dataset.

             Each line of this list contains the following information for each object:
             \itemize{
                 \item The label value.
                 \item The weight value.
                 \item The feature values.
             }
}
