% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{print.catboost.Pool}
\alias{print.catboost.Pool}
\title{Print catboost.Pool}
\usage{
\method{print}{catboost.Pool}(x, ...)
}
\arguments{
\item{x}{a catboost.Pool object

Default value: Required argument}

\item{...}{not currently used}
}
\value{
Nothing. This method prints pool dimensions.
}
\description{
Print dimensions of catboost.Pool.
}
