% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{summary.catboost.Model}
\alias{summary.catboost.Model}
\title{Print basic information about model}
\usage{
\method{summary}{catboost.Model}(object, ...)
}
\arguments{
\item{object}{The model obtained as the result of training.}

\item{...}{Not used}
}
\value{
The same model that was passed as input.
}
\description{
Displays the most general characteristics of a CatBoost model
(same as 'print').
}
