% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{tail.catboost.Pool}
\alias{tail.catboost.Pool}
\title{Tail of catboost.Pool}
\usage{
\method{tail}{catboost.Pool}(x, n = 10, ...)
}
\arguments{
\item{x}{The input dataset.

Default value: Required argument}

\item{n}{The quantity of the last objects in the dataset to be returned.

Default value: 10}

\item{...}{not currently used}
}
\description{
Return a list with the last n objects of the dataset.
}
\details{
Each line of this list contains the following information for each object:
\itemize{
    \item The target value.
    \item The weight value.
    \item The feature values.
}
}
