% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{catboost.cv}
\alias{catboost.cv}
\title{Cross-validate model.}
\usage{
catboost.cv(
  pool,
  params = list(),
  fold_count = 3,
  type = "Classical",
  partition_random_seed = 0,
  shuffle = TRUE,
  stratified = FALSE,
  early_stopping_rounds = NULL
)
}
\arguments{
\item{pool}{Data to cross-validate on}

\item{params}{Parameters for catboost.train}

\item{fold_count}{Folds count.}

\item{type}{is type of cross-validation.}

\item{partition_random_seed}{The random seed used for splitting pool into folds.}

\item{shuffle}{Shuffle the dataset objects before splitting into folds.}

\item{stratified}{Perform stratified sampling.}

\item{early_stopping_rounds}{Activates Iter overfitting detector with od_wait set to early_stopping_rounds.}
}
\value{
A data.frame of evaluation results from cross-validation.
}
\description{
Estimate model performance using cross-validation.
}
