% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{catboost.from_data_frame}
\alias{catboost.from_data_frame}
\title{catboost.from_data_frame}
\usage{
catboost.from_data_frame(data, target = NULL, weight = NULL,
  baseline = NULL, feature_names = NULL)
}
\arguments{
\item{data}{A data.frame with features.
The following column types are supported:
\itemize{
    \item double
    \item factor.
    It is assumed that categorical features are given in this type of columns.
    A standard CatBoost processing procedure is applied to this type of columns:
    \describe{
        \item{1.}{The values are converted to strings.}
        \item{2.}{The ConvertCatFeatureToFloat function is applied to the resulting string.}
    }
}

Default value: Required argument}

\item{target}{The target vector.}

\item{weight}{The weights of the target vector.}

\item{baseline}{Vector of initial (raw) values of the target function for the object.
Used in the calculation of final values of trees.}

\item{feature_names}{A list of names for each feature in the dataset.}
}
\value{
catboost.Pool
}
\description{
Create a dataset from the given data.frame.
Only numeric features are supported (their type should be double or factor).
Categorical features must be converted to numerical first.
For example, use \code{as.factor()} or the \code{colClasses} argument of the \code{read.table} method. The target type should be double.
}
\examples{
pool_path <- 'train_full3'
cd_vector <- c('numeric',  rep('numeric',2), rep('factor',7))
data <- read.table(pool_path, head = F, sep = "\\t", colClasses = cd_vector)
target <- c(1)
learn_size <- floor(0.8 * nrow(data))
learn_ind <- sample(nrow(data), learn_size)
learn <- data[learn_ind,]
test <- data[-learn_ind,]
learn_pool <- catboost.from_data_frame(data = learn[,-target], target = learn[,target])
test_pool <- catboost.from_data_frame(data = test[,-target], target = test[,target])

}
