% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{catboost.from_matrix}
\alias{catboost.from_matrix}
\title{Create pool from matrix}
\usage{
catboost.from_matrix(data, target, weight = NULL, baseline = NULL,
  cat_features = NULL)
}
\arguments{
\item{data}{Feature matrix}

\item{target}{Target vector}

\item{weight}{Weight vector}

\item{baseline}{Baseline vector}

\item{cat_features}{Vector of categorical feature indices}
}
\description{
Create pool from matrix
}
