% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{catboost.get_object_importance}
\alias{catboost.get_object_importance}
\title{Calculate the object importances}
\usage{
catboost.get_object_importance(
  model,
  pool,
  train_pool,
  top_size = -1,
  type = "Average",
  update_method = "SinglePoint",
  thread_count = -1,
  ostr_type = NULL
)
}
\arguments{
\item{model}{The model obtained as the result of training.

Default value: Required argument}

\item{pool}{The pool for which you want to evaluate the object importances.

Default value: Required argument}

\item{train_pool}{The pool on which the model has been trained.

Default value: Required argument}

\item{top_size}{Method returns the result of the top_size most important train objects. If -1, then the top size is not limited.

Default value: -1}

\item{type}{Possible values:
\itemize{
  \item 'Average'

    Method returns the mean train objects scores for all input objects.

  \item 'PerObject'

    Method returns the train objects scores for every input object.
}

Default value: 'Average'}

\item{update_method}{Description of the update set methods are given in section 3.1.3 of the paper.

Possible values:
\itemize{
  \item 'SinglePoint'
  \item 'TopKLeaves'
    It is posible to set top size : TopKLeaves:top=2.
  \item 'AllPoints'
}

Default value: 'SinglePoint'}

\item{thread_count}{The number of threads to use when applying the model. If -1, then the number of threads is set to the number of CPU cores.

Allows you to optimize the speed of execution. This parameter doesn't affect results.

Default value: -1}

\item{ostr_type}{Deprecated parameter, use 'type' instead.}
}
\value{
List with elements \code{"indices"} and \code{"scores"}.
}
\description{
Calculate the object importances (see \url{https://catboost.ai/docs/concepts/ostr.html}).
             This is the implementation of the LeafInfluence algorithm from the following paper:
              https://arxiv.org/pdf/1802.06640.pdf
}
\seealso{
\url{https://catboost.ai/docs/concepts/r-reference_catboost-get_object_importance.html}
}
