% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{catboost.importance}
\alias{catboost.importance}
\title{catboost.importance}
\usage{
catboost.importance(model, pool = NULL, thread_count = 1)
}
\arguments{
\item{model}{The model obtained as the result of training.

Default value: Required argument}

\item{pool}{The input dataset.

The feature importance for the training dataset is calculated if this argument is not specified.

Default value: NULL}

\item{thread_count}{The number of threads to use when applying the model.

Allows you to optimize the speed of execution. This parameter doesn't affect results.

Default value: 1}
}
\description{
Calculate the feature importances (see \url{https://tech.yandex.com/catboost/doc/dg/concepts/fstr-docpage/#fstr})
(Regular feature importance
(see \url{https://tech.yandex.com/catboost/doc/dg/concepts/output-data_feature-importance-docpage/#per-feature-importance})
and Feature interaction strength
(see \url{https://tech.yandex.com/catboost/doc/dg/concepts/output-data_feature-interaction-strength-docpage/#output-data_feature-interaction-strength})
).
}
\seealso{
\url{https://tech.yandex.com/catboost/doc/dg/concepts/r-reference_catboost-importance-docpage/}
}
