% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{catboost.load_model}
\alias{catboost.load_model}
\title{Load the model}
\usage{
catboost.load_model(model_path, file_format = "cbm")
}
\arguments{
\item{model_path}{The path to the model.

Default value: Required argument}

\item{file_format}{Format of the model file.

Default value: 'cbm'}
}
\value{
A model object.
}
\description{
Load the model from a file.

             Note: Feature importance (see \url{https://catboost.ai/docs/concepts/fstr.html#fstr})
             is not saved when using this function.
}
\seealso{
\url{https://catboost.ai/docs/concepts/r-reference_catboost-load_model.html}
}
