% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{catboost.ntrees}
\alias{catboost.ntrees}
\title{catboost.ntrees}
\usage{
catboost.ntrees(model)
}
\arguments{
\item{model}{The model obtained as the result of training.}
}
\value{
The number of trees.
}
\description{
Return the number of trees in the model.
}
\seealso{
\url{https://tech.yandex.com/catboost/doc/dg/concepts/r-reference_catboost-ntrees-docpage/}
}
