% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{catboost.predict}
\alias{catboost.predict}
\title{Get model prediction}
\usage{
catboost.predict(model, pool, verbose = FALSE, type = "RawFormulaVal",
  tree_count_limit = 0, thread_count = 1)
}
\arguments{
\item{model}{trained model}

\item{pool}{data to make predictions on}

\item{verbose}{output log while making prediction}

\item{type}{'Probability', 'Class' or 'RawFormulaVal'}

\item{tree_count_limit}{tree number to make prediction on}

\item{thread_count}{number of parallel threads}
}
\description{
Get model prediction
}
