% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{catboost.predict}
\alias{catboost.predict}
\alias{predict.catboost.Model}
\title{Get predictions from a CatBoost model}
\usage{
\method{predict}{catboost.Model}(
  object,
  newdata,
  verbose = FALSE,
  prediction_type = "RawFormulaVal",
  ntree_start = 0,
  ntree_end = 0,
  thread_count = -1
)

catboost.predict(
  object,
  newdata,
  verbose = FALSE,
  prediction_type = "RawFormulaVal",
  ntree_start = 0,
  ntree_end = 0,
  thread_count = -1
)
}
\arguments{
\item{object}{The model obtained as the result of training.

Default value: Required argument}

\item{newdata}{The input data on which to make predictions. Should be a `catboost.Pool`
object.

Default value: Required argument}

\item{verbose}{Verbose output to stdout.

Default value: FALSE (not used)}

\item{prediction_type}{The format for displaying approximated values in output data
(see \url{https://catboost.ai/docs/concepts/output-data.html}).

Possible values:
\itemize{
  \item 'Probability'
  \item 'LogProbability'
  \item 'Class'
  \item 'RawFormulaVal'
  \item 'Exponent'
  \item 'RMSEWithUncertainty'
}

Default value: 'RawFormulaVal'}

\item{ntree_start}{Model is applied on the interval [ntree_start, ntree_end) (zero-based indexing).

Default value: 0}

\item{ntree_end}{Model is applied on the interval [ntree_start, ntree_end) (zero-based indexing).

Default value: 0 (if value equals to 0 this parameter is ignored and ntree_end equal to tree_count)}

\item{thread_count}{The number of threads to use when applying the model. If -1, then the number of threads is set to the number of CPU cores.

Allows you to optimize the speed of execution. This parameter doesn't affect results.

Default value: 1}
}
\value{
Vector of predictions (matrix for multi-class classification).
}
\description{
Get predictions from a CatBoost model on new data.
}
\details{
The function `catboost.predict` is a synonym for `predict.catboost.Model`,
which is an S3 method (i.e. called like `predict(model, newdata)`).

In case of multiclassification the prediction is returned in the form of a matrix.
Each row of this matrix contains the predictions for one row of the input dataset.
}
\seealso{
\url{https://catboost.ai/docs/concepts/r-reference_catboost-predict.html}
}
