% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{catboost.save_model}
\alias{catboost.save_model}
\title{Save the model}
\usage{
catboost.save_model(
  model,
  model_path,
  file_format = "cbm",
  export_parameters = NULL,
  pool = NULL
)
}
\arguments{
\item{model}{The model to be saved.

Default value: Required argument}

\item{model_path}{The path to the resulting binary file with the model description.
Used for solving other machine learning problems (for instance, applying a model).

Default value: Required argument}

\item{file_format}{specified format model from a file.
Possible values:
\itemize{
  \item 'cbm'
    For catboost binary format
  \item 'coreml'
    To export into Apple CoreML format
  \item 'onnx'
    To export into ONNX-ML format
  \item 'pmml'
    To export into PMML format
  \item 'cpp'
    To export as C++ code
  \item 'python'
    To export as Python code.
}

Default value: 'cbm'}

\item{export_parameters}{are a parameters for CoreML or PMML export.}

\item{pool}{is training pool.}
}
\value{
Status, the result of model shrinking. TRUE if shrinking succeeded, FALSE otherwise.
}
\description{
Save the model to a file.

             Note: Feature importance (see \url{https://catboost.ai/docs/concepts/fstr.html#fstr})
             is not saved when using this function.
}
\seealso{
\url{https://catboost.ai/docs/features/export-model-to-core-ml.html}
}
