% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{catboost.save_pool}
\alias{catboost.save_pool}
\title{Save the dataset}
\usage{
catboost.save_pool(data, label = NULL, weight = NULL,
  baseline = NULL, pool_path = "data.pool", cd_path = "cd.pool")
}
\arguments{
\item{data}{A data.frame with features.
The following column types are supported:
    \itemize{
    \item double
    \item factor.
    It is assumed that categorical features are given in this type of columns.
    A standard CatBoost processing procedure is applied to this type of columns:
    \describe{
        \item{1.}{The values are converted to strings.}
        \item{2.}{The ConvertCatFeatureToFloat function is applied to the resulting string.}
    }
}

Default value: Required argument}

\item{label}{The label vector.}

\item{weight}{The weights of the label vector.}

\item{baseline}{Vector of initial (raw) values of the label function for the object.
Used in the calculation of final values of trees.}

\item{pool_path}{The path to the otuptut file that contains the dataset description.}

\item{cd_path}{The path to the output file that contains the column descriptions.}
}
\description{
Save the dataset to the CatBoost format.
Files with the following data are created:
\itemize{
    \item Dataset description
    \item Column descriptions
}
Use the catboost.load_pool function to read the resulting files.
These files can also be used in the \href{https://catboost.ai/docs/concepts/cli-installation.html}{Command-line version}
and the \href{https://catboost.ai/docs/concepts/python-installation.html}{Python library}.
}
